requirements/9MAR2021.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.3
command-not-found==0.3
compizconfig-python==0.9.14.1
cryptography==3.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
distro-info===0.23ubuntu1
Django==3.1.5
django-session-timeout
django-taggit==1.2.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
gyp==0.1
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.17.0
jeepney==0.4.3
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
keyring==21.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
libevdev==0.5
lockfile==0.12.2
louis==3.14.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
meson==0.55.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.6
protobuf==3.12.3
psutil==5.7.2
pycairo==1.16.2
pycups==2.0.1
Pygments==2.3.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.1.3+ubuntu1.3
python-dateutil==2.8.1
python-debian==0.1.37
python-distutils-extra==2.39
python-Levenshtein==0.12.0
python-xapp==2.0.1
pytz==2020.1
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
reportlab==3.5.47
requests==2.23.0
requests-unixsocket==0.2.0
scour==0.37
SecretStorage==3.1.2
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.17.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
terminado==0.8.2
testpath==0.4.4
texttable==1.6.2
tornado==6.0.4
traitlets==4.3.3
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
ulauncher==5.10.0
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
virtualenv==20.0.29+ds
wadllib==1.3.4
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.53.0
widgetsnbextension==2.0.0
xkit==0.0.0
zipp==1.0.0
